chore: fix build problems
