+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/autocrafting/CraftingPattern.java
@@ -1,15 +1,17 @@
 package com.refinedmods.refinedstorage.common.autocrafting;
 
-import com.refinedmods.refinedstorage.api.autocrafting.AbstractPattern;
+import com.refinedmods.refinedstorage.api.autocrafting.Pattern;
 import com.refinedmods.refinedstorage.api.resource.ResourceAmount;
 import com.refinedmods.refinedstorage.api.resource.ResourceKey;
 import com.refinedmods.refinedstorage.common.api.support.resource.PlatformResourceKey;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Set;
 import java.util.UUID;
 
-class CraftingPattern extends AbstractPattern {
+class CraftingPattern implements Pattern {
+    private final UUID id;
     private final List<List<PlatformResourceKey>> inputs;
     private final ResourceAmount output;
     private final List<ResourceAmount> byproducts;
@@ -19,7 +21,7 @@ class CraftingPattern extends AbstractPattern {
                     final List<List<PlatformResourceKey>> inputs,
                     final ResourceAmount output,
                     final List<ResourceAmount> byproducts) {
-        super(id);
+        this.id = id;
         this.inputs = inputs;
         this.output = output;
         this.outputResources = Set.of(output.resource());
@@ -37,5 +39,22 @@ class CraftingPattern extends AbstractPattern {
 
     ResourceAmount getOutput() {
         return output;
+    }
+
+    @Override
+    public boolean equals(final Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        final CraftingPattern that = (CraftingPattern) o;
+        return Objects.equals(id, that.id);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hashCode(id);
     }
 }
