+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/autocrafting/ProcessingPattern.java
@@ -1,21 +1,23 @@
 package com.refinedmods.refinedstorage.common.autocrafting;
 
-import com.refinedmods.refinedstorage.api.autocrafting.AbstractPattern;
+import com.refinedmods.refinedstorage.api.autocrafting.Pattern;
 import com.refinedmods.refinedstorage.api.resource.ResourceAmount;
 import com.refinedmods.refinedstorage.api.resource.ResourceKey;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Set;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
-class ProcessingPattern extends AbstractPattern {
+class ProcessingPattern implements Pattern {
+    private final UUID id;
     private final List<ResourceAmount> inputs;
     private final List<ResourceAmount> outputs;
     private final Set<ResourceKey> outputResources;
 
     ProcessingPattern(final UUID id, final List<ResourceAmount> inputs, final List<ResourceAmount> outputs) {
-        super(id);
+        this.id = id;
         this.inputs = inputs;
         this.outputs = outputs;
         this.outputResources = outputs.stream().map(ResourceAmount::resource).collect(Collectors.toSet());
@@ -26,11 +28,24 @@ class ProcessingPattern extends AbstractPattern {
         return outputResources;
     }
 
-    List<ResourceAmount> getInputs() {
-        return inputs;
-    }
-
     List<ResourceAmount> getOutputs() {
         return outputs;
+    }
+
+    @Override
+    public boolean equals(final Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        final ProcessingPattern that = (ProcessingPattern) o;
+        return Objects.equals(id, that.id);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hashCode(id);
     }
 }
